refactor: placeholder.js 파일 constants 폴더로 이동
