+++ src/constants/placeholder.js
